Bail out if recipient is unknown
So far, the recipient was evaluated after the message was given. That is
not how o2sms behaved and I find it very unpleasant to have to search
and retype the message if the alias is wrong, i.e. mistyped. So what we
do now is to raise an error and print the available aliases.

Conflicts:
	cliresms