Enabled config to contain group definitions, i.e. alias foo bar baz
That was a nice feature of o2sms. You could define
alais beerpeople greg 1234 bob
and write to both of them easily
That was currently not supported. Now it should work.

This is based on work done by Tobias Mueller. I just added parsing
using regexes to allow for the parsing to be a bit more flexible and
resilient to slightly malformed but otherwise correct lines.